Please see these for the interactive maps.
Please note that, due to quarto render memory restriction and Github 25mb file size restriction, I'm not able to upload some of the generated visulaizations to the website. Therefore, I'm inserting screenshots in the website to show the results that I ran in the Jupyter notebook. Please make sure to open the interactive html files here to navigate the walkability analysis part